migrated read tests
